+++ src/Models/Volumes/VolumeRequestOpts.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: lukaskammerling
+ * Date: 2019-03-28
+ * Time: 13:51
+ */
+
+namespace LKDev\HetznerCloud\Models\Volumes;
+
+
+use LKDev\HetznerCloud\RequestOpts;
+
+class VolumeRequestOpts extends RequestOpts
+{
+    /**
+     * @var string
+     */
+    public $name;
+
+    /**
+     * @var string
+     */
+    public $status;
+
+    /**
+     * RequestOpts constructor.
+     *
+     * @param $name
+     * @param $perPage
+     * @param $page
+     * @param $labelSelector
+     */
+    public function __construct(string $name = null, string $status = null, int $perPage = null, int $page = null, string $labelSelector = null)
+    {
+        parent::__construct($perPage, $page, $labelSelector);
+        $this->name = $name;
+        $this->status = $status;
+    }
+}